refactor: add hooks/data subfolder for useData use cases
