+++ src/hooks/data/useDeploymentStrategiesData.ts
@@ -1,0 +1,8 @@
+import { getDeploymentStrategies } from '../../api/deployment-strategies';
+import { deploymentStrategiesAtom } from '../../atoms/deployment-strategies';
+import { useData } from '../useData';
+
+export const useDeploymentStrategiesData = () =>
+    useData(deploymentStrategiesAtom, getDeploymentStrategies, {
+        useRawType: true,
+    });
