+++ src/hooks/data/useLibrariesData.ts
@@ -1,0 +1,6 @@
+import { getLibraries } from '../../api/library';
+import { librariesAtom } from '../../atoms/library';
+import { useData } from '../useData';
+
+export const useLibrariesData = () =>
+    useData(librariesAtom, getLibraries, { useRawType: true });
