+++ src/hooks/data/useModelMetadata.ts
@@ -1,0 +1,6 @@
+import { getModelMetadataById } from '../../api/models';
+import { modelMetadataAtom } from '../../atoms/model-metadata';
+import { useData } from '../useData';
+
+export const useModelMetadata = () =>
+    useData(modelMetadataAtom, getModelMetadataById);
