--- src/hooks/useDeploymentStrategiesData.ts
@@ -1,8 +1,0 @@
-import { getDeploymentStrategies } from '../api/deployment-strategies';
-import { deploymentStrategiesAtom } from '../atoms/deployment-strategies';
-import { useData } from './useData';
-
-export const useDeploymentStrategiesData = () =>
-    useData(deploymentStrategiesAtom, getDeploymentStrategies, {
-        useRawType: true,
-    });
