--- src/hooks/useLibrariesData.ts
@@ -1,6 +1,0 @@
-import { getLibraries } from '../api/library';
-import { librariesAtom } from '../atoms/library';
-import { useData } from './useData';
-
-export const useLibrariesData = () =>
-    useData(librariesAtom, getLibraries, { useRawType: true });
