--- src/hooks/useModelsData.ts
@@ -1,5 +1,0 @@
-import getModels from '../api/models';
-import { modelsAtom } from '../atoms/models';
-import { useData } from './useData';
-
-export const useModelsData = () => useData(modelsAtom, getModels);
