--- src/hooks/useResourcesData.ts
@@ -1,6 +1,0 @@
-import { getResources } from '../api/resources';
-import { resourcesAtom } from '../atoms/resources';
-import { useData } from './useData';
-
-export const useResourcesData = () =>
-    useData(resourcesAtom, getResources, { useRawType: true });
